+++ app/src/main/java/com/codeest/geeknews/model/event/SearchEvent.java
@@ -1,4 +1,4 @@
-package com.codeest.geeknews.model.bean;
+package com.codeest.geeknews.model.event;
 
 /**
  * Created by codeest on 16/8/31.
