+++ app/src/main/java/com/codeest/geeknews/model/http/api/WeChatApis.java
@@ -1,6 +1,7 @@
-package com.codeest.geeknews.model.http;
+package com.codeest.geeknews.model.http.api;
 
 import com.codeest.geeknews.model.bean.WXItemBean;
+import com.codeest.geeknews.model.http.response.WXHttpResponse;
 
 import java.util.List;
