+++ app/src/main/java/com/codeest/geeknews/model/http/response/GankHttpResponse.java
@@ -1,4 +1,4 @@
-package com.codeest.geeknews.model.http;
+package com.codeest.geeknews.model.http.response;
 
 /**
  * Created by codeest on 2016/8/3.
