+++ app/src/main/java/com/codeest/geeknews/presenter/TechPresenter.java
@@ -5,8 +5,8 @@ import com.codeest.geeknews.base.RxPresenter;
 import com.codeest.geeknews.component.RxBus;
 import com.codeest.geeknews.model.bean.GankItemBean;
 import com.codeest.geeknews.model.bean.GankSearchItemBean;
-import com.codeest.geeknews.model.bean.SearchEvent;
-import com.codeest.geeknews.model.http.GankHttpResponse;
+import com.codeest.geeknews.model.event.SearchEvent;
+import com.codeest.geeknews.model.http.response.GankHttpResponse;
 import com.codeest.geeknews.model.http.RetrofitHelper;
 import com.codeest.geeknews.presenter.contract.TechContract;
 import com.codeest.geeknews.ui.gank.fragment.GankMainFragment;
