+++ app/src/main/java/com/codeest/geeknews/ui/main/activity/MainActivity.java
@@ -19,7 +19,7 @@ import com.codeest.geeknews.app.Constants;
 import com.codeest.geeknews.base.BaseActivity;
 import com.codeest.geeknews.component.RxBus;
 import com.codeest.geeknews.component.UpdateService;
-import com.codeest.geeknews.model.bean.SearchEvent;
+import com.codeest.geeknews.model.event.SearchEvent;
 import com.codeest.geeknews.presenter.MainPresenter;
 import com.codeest.geeknews.presenter.contract.MainContract;
 import com.codeest.geeknews.ui.gank.fragment.GankMainFragment;
